Add files via upload
Greeting, Students!

This is an algorithm for Predictive Analysis of Boston Housing Prices Using PySpark.

It implements a predictive analysis of housing prices in Boston using the PySpark framework, focusing on determining which factors most strongly predict the median value of homes. It starts by initializing a Spark session and setting up a schema for the data, which is then loaded into a DataFrame from a CSV file. The script calculates pairwise correlations between the median value of homes ('MEDV') and other variables in the dataset, identifying the top three most influential factors. These variables are used to create polynomial features, which are then employed in a linear regression model to predict housing prices. The model is trained on a split of 70% of the data, and its performance is evaluated on the remaining 30% using the R-squared metric to measure the accuracy of the predictions. The implementation uses a pipeline to streamline the process of vectorizing features, transforming them into polynomial space, and applying linear regression.

Happy Learning,
Farshid Keivanian
https://www.linkedin.com/in/farshid-keivanian/
https://keivan-rc.com/